feat: 【i】【优化】【代码】

+++ be-base-20240901/be-base-web-20240901/src/main/java/com/kar20240901/be/base/web/model/constant/base/BaseConstant.java
@@ -7,6 +7,6 @@ public interface BaseConstant {
 
     String TOP_FOLDER_NAME = "根文件夹";
 
-    int FILE_CHUNK_SIZE = 1024 * 1024 * 5; // 每个分片文件最大大小
+    Long FILE_CHUNK_SIZE = 1024 * 1024 * 5L; // 每个分片文件最大大小
 
 }
