+++ be-base-20240901/be-base-web-20240901/src/main/java/com/kar20240901/be/base/web/model/domain/file/BaseFileTransferDO.java
@@ -45,7 +45,7 @@ public class BaseFileTransferDO extends TempEntity {
     private String fileSign;
 
     @Schema(description = "每个分片的大小")
-    private Integer chunkSize;
+    private Long chunkSize;
 
     @Schema(description = "总分片个数")
     private Integer chunkTotal;
