+++ be-base-20240901/be-base-web-20240901/src/main/java/com/kar20240901/be/base/web/model/vo/file/BaseFileUploadChunkPreVO.java
@@ -10,7 +10,7 @@ public class BaseFileUploadChunkPreVO {
     private Long fileId;
 
     @Schema(description = "每个分片的大小")
-    private Integer chunkSize;
+    private Long chunkSize;
 
     @Schema(description = "总分片个数")
     private Integer chunkTotal;
